Double Dash Added
The player can now perform a secondary dash when hitting an opponent.
The second Dash does not deal damage.
Because of how Dash reset works,  resetting the dash when touching the ground.
double dashes must be used in mid air, or within a extremely short period after if on ground.

There is currently no visual indication for when secondary dash is available.

Used level 1(Dubbed the debug level) to test, added 2 flying slimes to test.

Requires further testing.