Fixed checkpoints
Fixed checkpoints to respawn the player at the location of the checkpoint instead of at the start of the level.

Ranged slime projectile has been moved down to let the player jump over it easier.

Player jump has also been increased to allow the player to groundpound slimes.

Both the game and the editor now open on the main menu instead of level 1.